refactor: combine `Client` and `Application` into `BasicWDClient`
